+++ finalMVA/mkNtuple/makeTuple.h
@@ -96,7 +96,7 @@ public :
    virtual TList  *GetOutputList() const { return fOutput; }
    virtual void    SlaveTerminate();
    virtual void    Terminate();
-   double transverseMass(const TLorentzVector & l, const TLorentzVector & nu);
+   float transverseMass(const TLorentzVector & l, const TLorentzVector & nu);
    //int totalevt = fChain->GetTree()->GetEntries();
 
     TTree *tree = 0;
